elemet 추가 이슈 해결, 로그인 페이지 구현 및 SPA 적용
