+++ front_end/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <title>Ajou Bulletin Board</title>
+    <link rel="stylesheet" href="index.css" />
+</head>
+
+<body>
+    <div class="menu-bar">
+        <div id="sliding-menu-button"><img src="menu_button.svg" width="15px" height="15px"></div>
+        <div id="main-button" data-link="">Ajou Memo</div>
+        <div id="login-button" data-link="login">Log in</div>
+    </div>
+
+    <div class="root-content">
+    </div>
+
+    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
+    <script type="module" src="./index.js"></script>
+</body>
+
+</html>
