--- front_end/page.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <title>Ajou Bulletin Board</title>
-    <link rel="stylesheet" href="page.css" />
-</head>
-
-<body>
-    <div class="menu-bar">
-        <div class="sliding-menu-button"><img src="menu_button.svg" width="15px" height="15px"></div>
-        <div>Ajou Memo</div>
-        <div>Log in</div>
-    </div>
-
-    <input type="button" value="글쓰기" data-link="write-post" />
-
-    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-    <script type="module" src="page.js"></script>
-</body>
-
-</html>
